Fix default FILEPATH for cpct_winape in WSL
When using WSL, since wslpath uses the pwd command, when no file is given to cpct_winape, the working directory is passed to WinAPE.exe as a the file to run. This causes an error when closing WinAPE, since it tries to save a nonexistent file.

I fixed this by only setting the FILEPATH variable when the FILE variable is set (in other words, when a file is given to cpct_winape).